WICKET-5254
First draft of implementation, however one test is failing, so in its current state it is broken.
The idea behind is set an application meta data property allowing to tell if we want to prevent modifying accessibility of a field / method.
The bad part is that we change accessibility of fields or methods during getter/setter lookup, thus, throwing a runtimeexception is overkill in this situation (that's why PropertyResolverTest#propertyModel is failing).